feat: OOP
- In order for the data to be transfered between filed I updated files to use OOP structure
